Add hcl scheme validation for 'inline' configurations #59
